refactor(core): move loadDefaultWorkflows into @bakin/core/workflows
The images plugin registered its shipped workflow defaults through
plugins/workflows/lib/load-defaults — a plugin-to-plugin import. Move
the loader (body byte-for-byte, YAML parse inlined via js-yaml, which
core already uses) to packages/core/src/workflows/load-defaults.ts
next to validateDefinition, exactly like the 5cf27fe6 move. The
workflows plugin's module becomes a re-export shim so plugin-internal
consumers and tests are untouched; images imports the core home.
