+++ plugins/images/index.ts
@@ -7,7 +7,7 @@ import { z } from 'zod'
 import type { BakinPlugin, ExecToolResult, HealthCheckResult, PluginContext } from '@bakin/core/plugin-types'
 import { definePlugin, defineRoute } from '@bakin/core/routing'
 import { createLogger } from '../../src/core/logger'
-import { loadDefaultWorkflows } from '../workflows/lib/load-defaults'
+import { loadDefaultWorkflows } from '@bakin/core/workflows/load-defaults'
 import { DEFAULT_IMAGE_SETTINGS, listImageProviders, providerReadiness } from './lib/providers'
 import { getImageProfile, listImageProfiles } from './lib/platform-profiles'
 import { editImage, exportImage, generateImage, importImage } from './lib/tools'
